+++ src/main/java/org/kitteh/irc/client/library/defaults/DefaultClient.java
@@ -45,9 +45,9 @@ import org.kitteh.irc.client.library.element.User;
 import org.kitteh.irc.client.library.element.mode.ModeStatus;
 import org.kitteh.irc.client.library.element.mode.ModeStatusList;
 import org.kitteh.irc.client.library.element.mode.UserMode;
-import org.kitteh.irc.client.library.event.client.ClientBatchEndEvent;
-import org.kitteh.irc.client.library.event.client.ClientBatchMessageEvent;
-import org.kitteh.irc.client.library.event.client.ClientBatchStartEvent;
+import org.kitteh.irc.client.library.event.batch.ClientBatchEndEvent;
+import org.kitteh.irc.client.library.event.batch.ClientBatchMessageEvent;
+import org.kitteh.irc.client.library.event.batch.ClientBatchStartEvent;
 import org.kitteh.irc.client.library.event.client.ClientReceiveCommandEvent;
 import org.kitteh.irc.client.library.event.client.ClientReceiveNumericEvent;
 import org.kitteh.irc.client.library.event.helper.ClientReceiveServerMessageEvent;
