+++ src/main/java/org/kitteh/irc/client/library/feature/sts/StsHandler.java
@@ -29,7 +29,7 @@ import org.kitteh.irc.client.library.element.CapabilityState;
 import org.kitteh.irc.client.library.element.ServerMessage;
 import org.kitteh.irc.client.library.event.capabilities.CapabilitiesNewSupportedEvent;
 import org.kitteh.irc.client.library.event.capabilities.CapabilitiesSupportedListEvent;
-import org.kitteh.irc.client.library.event.client.ClientConnectionEndedEvent;
+import org.kitteh.irc.client.library.event.connection.ClientConnectionEndedEvent;
 import org.kitteh.irc.client.library.exception.KittehServerMessageException;
 import org.kitteh.irc.client.library.util.HostWithPort;
 import org.kitteh.irc.client.library.util.StsUtil;
